requirements.txt
python-telegram-bot==21.4
configparser==7.0.0